names: move 9 forms to Views/
Per #4331. Mechanical move of all 9 single-use Name forms from
`app/components/` to `app/views/controllers/names/...`:

- `NameForm` → `Views::Controllers::Names::Form`
- `NameClassificationForm` →
  `Views::Controllers::Names::Classification::Form`
- `NameInheritClassificationForm` →
  `Views::Controllers::Names::Classification::Inherit::Form`
- `NameLifeformForm` →
  `Views::Controllers::Names::Lifeforms::Form`
- `NamePropagateLifeformForm` →
  `Views::Controllers::Names::Lifeforms::Propagate::Form`
- `NameApproveSynonymForm` →
  `Views::Controllers::Names::Synonyms::Approve::Form`
- `NameDeprecateSynonymForm` →
  `Views::Controllers::Names::Synonyms::Deprecate::Form`
- `NameEditSynonymForm` →
  `Views::Controllers::Names::Synonyms::Form`
- `NameTrackerForm` →
  `Views::Controllers::Names::Trackers::Form`

Each new view's namespace mirrors its controller path. Tests moved
+ namespaced to match.

The form-id auto-deriver in `ApplicationForm` would have produced a
shorter id (e.g. `lifeform_form` instead of `name_lifeform_form`)
because it only uses the parent module of the form's class. The
names cluster is special — the old class names mashed "Name" into
the form name (`NameLifeformForm` → `name_lifeform_form`) — so each
moved form passes its old id explicitly via `super(..., id: "...")`
to preserve the contract for any downstream Stimulus / CSS hooks.
Going forward, new forms get the auto-derived id naturally.

The 6 FormObject-bound forms (synonym ops, lifeform string↔checkbox
translation, classification-inherit lookup, propagation params) all
correctly represent UI-level operations distinct from Name
attributes — not refactor candidates.

291 tests pass; lint clean.
