+++ app/views/controllers/names/classification/edit.html.erb
@@ -6,4 +6,4 @@ add_context_nav(name_forms_return_tabs(name: @name))
 container_class(:text)
 %>
 
-<%= render(Components::NameClassificationForm.new(@name)) %>
+<%= render(Views::Controllers::Names::Classification::Form.new(@name)) %>
