+++ app/views/controllers/names/edit.html.erb
@@ -3,7 +3,7 @@ add_edit_title(@name, user: @user)
 add_context_nav(name_form_edit_tabs(name: @name))
 %>
 
-<%= render(Components::NameForm.new(@name, user: @user,
+<%= render(Views::Controllers::Names::Form.new(@name, user: @user,
                                     name_string: @name_string,
                                     misspelling: @misspelling,
                                     correct_spelling: @correct_spelling)) %>
