+++ app/views/controllers/names/lifeforms/edit.html.erb
@@ -5,7 +5,7 @@ add_context_nav(name_forms_return_tabs(name: @name))
 container_class(:text_image)
 %>
 
-<%= render(Components::NameLifeformForm.new(
+<%= render(Views::Controllers::Names::Lifeforms::Form.new(
       FormObject::Lifeform.from_name(@name),
       name: @name
     )) %>
