+++ app/views/controllers/names/new.html.erb
@@ -3,6 +3,6 @@ add_new_title(:create_object, :NAME)
 add_context_nav(name_form_new_tabs)
 %>
 
-<%= render(Components::NameForm.new(@name, user: @user,
+<%= render(Views::Controllers::Names::Form.new(@name, user: @user,
                                     name_string: @name_string,
                                     approved_rank: @approved_rank)) %>
