+++ app/views/controllers/names/synonyms/approve/new.html.erb
@@ -4,7 +4,7 @@ add_page_title(:name_approve_title.t(name: @name.display_name))
 add_context_nav(name_forms_return_tabs(name: @name))
 %>
 
-<%= render(Components::NameApproveSynonymForm.new(
+<%= render(Views::Controllers::Names::Synonyms::Approve::Form.new(
       FormObject::ApproveSynonym.new(deprecate_others: true),
       name: @name,
       approved_names: @approved_names
