+++ app/views/controllers/names/synonyms/deprecate/new.html.erb
@@ -4,7 +4,7 @@ add_page_title(:name_deprecate_title.t(name: @name.display_name))
 add_context_nav(name_forms_return_tabs(name: @name))
 %>
 
-<%= render(Components::NameDeprecateSynonymForm.new(
+<%= render(Views::Controllers::Names::Synonyms::Deprecate::Form.new(
       name: @name,
       proposed_name: @given_name,
       is_misspelling: @misspelling,
