+++ app/views/controllers/names/trackers/new.html.erb
@@ -13,7 +13,7 @@ add_context_nav(name_forms_return_tabs(name: @name))
   end %>
 </div>
 
-<%= render(Components::NameTrackerForm.new(
+<%= render(Views::Controllers::Names::Trackers::Form.new(
       NameTracker.new(name: @name),
       note_template: @note_template
     )) %>
